backend/api/mixins.py
from rest_framework import mixins, viewsets


class CreateListRetrieveMixin(
    mixins.CreateModelMixin,
    mixins.ListModelMixin,
    mixins.RetrieveModelMixin,
    viewsets.GenericViewSet
):
    pass
