backend/requirements.txt
django==4.2.10
django-filter==23.5
djangorestframework==3.14.0
djoser==2.2.2